@@ -8,7 +8,7 @@ import indotech.io.mybatis.model.response.PaginationDataResponse;
 import java.math.BigDecimal;
 
 public interface CompanyService {
-    PaginationDataResponse<CompanyDTO> getDataWithPagination(int page, int limit, String sortField, String sortOrder);
+    PaginationDataResponse<CompanyDTO> getDataWithPagination(int page, int limit, String orderField, String orderType);
     DataResponse<CompanyDTO> findOne(BigDecimal id);
     DataResponse<CompanyDTO> create(CompanyDTO company);
     DataResponse<CompanyDTO> update(BigDecimal id, CompanyDTO company);
